add function to build specific gpu-arches [REVIEW] (#766)
This PR adds in functions to the root level build.sh so that you can specify which GPU architectures you want to build for. Currently the only options were to build for the current system or for all systems. It also adds this option into `examples/build.sh` in addition to fixing bugs with the clean command.

A `--gpu-arch` option is added that is meant to be used to provide which architectures to build for e.g. `./build.sh bench-ann --gpu-arch="80-real;90-real"`

Authors:
  - Rohan Varma (https://github.com/nvrohanv)
  - Gil Forsyth (https://github.com/gforsyth)

Approvers:
  - Ben Karsin (https://github.com/bkarsin)
  - Corey J. Nolet (https://github.com/cjnolet)
  - Micka (https://github.com/lowener)
  - Gil Forsyth (https://github.com/gforsyth)

URL: https://github.com/rapidsai/cuvs/pull/766