Moving NN Descent class and struct declarations to `nn_descent_gnnd.hpp` (#803)
Moves declaration of the `GNND` class and a few other classes and structs to `nn_descent_gnnd.hpp` file.
Also adds a helper function for getting the build configurations based on a few checks.

## Why this is needed
Actual work for NN Descent build is done in the `GNND::build()` function, which is not exposed as a public API.
This complicates usage of NN Descent, especially when we want to create the GNND separately, and call gnnd.build() directly.

## Use Cases
Specifically, this is needed for the batched all-neighbors knn graph building implementation. 
GNND allocates all necessary device and host matrices at initialization.
The NN Descent's `build` exposed as the public API makes a new GNND, then calls gnnd.build().
So originally it's like this
```
public api build() {
     gnnd = GNND{};   // allocates a bunch of device/host memory
     gnnd.build(data);    // where all the work is done
}
```
However, now that we have to build iteratively in the batching algorithm, calling the public build API means we make GNND for every new cluster, thereby allocating (and deallocating) a large chunk of memory for every cluster. This takes a lot of time.

In such cases, it is better to make a GNND once, and reuse it over multiple clusters to save time. For example;
```
gnnd = GNND{}    // this needs to be made outside the for loop
for cluster in clusters:
    cluster_data = get_cluster_data(data)
    gnnd.build(cluster_data)
   ```
This PR will allow including `nn_descent_gnnd.hpp` instead of the entire `nn_descent.cuh` file.

Authors:
  - Jinsol Park (https://github.com/jinsolp)

Approvers:
  - Divye Gala (https://github.com/divyegala)

URL: https://github.com/rapidsai/cuvs/pull/803