[CuVS-Java] Automate panama bindings generation, Include IVF_PQ parameters in CAGRA index parameters and other changes (#831)
The PR includes code changes for the following:

- Automation of Panama bindings generation using jextract.
- Adding the ability to configure IVF_PQ index and search parameters via the cuvs-java API (to adapt with the following [underlying changes](https://github.com/rapidsai/cuvs/commit/a14efd5a10502962a2a6391e6063f8ab688ddff1)). 
- Updating HNSW example to show the above.
- Updating the readme files.
- Simplifying logging in examples.
- Bumping up the maven-javadoc-plugin version.
- Updating and consolidating gitignore file.
- Removing unused imports etc.

Please note that the existing Panama classes are being deleted because they were manually created and managed. With the new cleaner approach, this will not be needed anymore. Now these binding classes will be generated at build time and so no need to be in the codebase.

Authors:
  - Vivek Narang (https://github.com/narangvivek10)
  - Ishan Chattopadhyaya (https://github.com/chatman)

Approvers:
  - rhdong (https://github.com/rhdong)
  - James Lamb (https://github.com/jameslamb)

URL: https://github.com/rapidsai/cuvs/pull/831