batch_load_iterator shall copy data located in host memory (#926)
This PR changes the `batch_load_iterator` helper to copy the data explicitly if it is host accessible.

Previously, if the data was both host and device accessible, then there was no explicit copy done by `batch_load_iterator`, since we could access the data directly on the GPU.

In practice, the datasets handled by `batch_load_iterator` are expected to be large, and in that case relying on automatic page migration to access the data is not as efficient as the explicit copy facilitated by `batch_load_iterator`. This PR enables the explicit copy op by `batch_load_iterator` in this case.

Authors:
  - Tamas Bela Feher (https://github.com/tfeher)

Approvers:
  - Artem M. Chirkin (https://github.com/achirkin)

URL: https://github.com/rapidsai/cuvs/pull/926