[Java] Add Java API benchmarks (#1033)
This PR introduces a maven project configured with JMH, so we can write (micro-)benchmarks to test out different implementation choices and alternatives and their relative performances.

As an example, the PR introduces 2 benchmarks for the Cagra index API: one testing the performances of building and serializing, the other the performances of building an index from a dataset in heap memory. I plan to follow up with PRs that introduce different ways to serialize an index, and different way of specify an input dataset (off-heap memory), adding corresponding tests to see what are the performance gains.

Authors:
  - Lorenzo Dematté (https://github.com/ldematte)
  - Ben Frederickson (https://github.com/benfred)
  - MithunR (https://github.com/mythrocks)

Approvers:
  - Chris Hegarty (https://github.com/ChrisHegarty)
  - MithunR (https://github.com/mythrocks)

URL: https://github.com/rapidsai/cuvs/pull/1033