[Java] Introduce scoped resource access (#1089)
This PR wraps access to the cuVS resources handle into a autocloseable accessor. The reason is to provide a clear scope for resource access, with the possibility to add application logic to the scope begin/end.

An example of this logic is the `CheckedCuVSResources` decorator introduced in the test suite, which checks that the `CuVSResources` it decorates has not been destroyed yet, and that access to the underlying cuVS resources handle is not concurrent (is not done by more than 1 thread at every time).

I preferred to write it as a decorator, as this gives the API consumers more freedom: if they want the check, they can use `CheckedCuVSResources` (or a similar class); if they want direct, straight, fast access, they can just use the accessor; if they need more logic (e.g. logging or anything else), that is possible too.

Tests are updated to use  `CheckedCuVSResources`, so we can spot issues quickly in tests without burdening production code with additional logic.

Authors:
  - Lorenzo Dematté (https://github.com/ldematte)
  - MithunR (https://github.com/mythrocks)

Approvers:
  - MithunR (https://github.com/mythrocks)

URL: https://github.com/rapidsai/cuvs/pull/1089