[Java] Exception-safe RMM Allocations (#1215)
This commit introduces exception safety for RMM allocations.

Previously, device memory allocated through `cuvsRmmAlloc()` was freed manually using `cuvsRmmFree()`, in all the index impl classes.  The problem there is that if an exception is thrown in the intervening time between alloc and free, it would lead to a leak of device memory.

This commit extends the `CloseableHandle` class to encapsulate the allocation of device memory.  This new class is used in try-with-resources blocks, to make device memory allocations exception-safe.

Authors:
  - MithunR (https://github.com/mythrocks)

Approvers:
  - Lorenzo Dematté (https://github.com/ldematte)
  - Corey J. Nolet (https://github.com/cjnolet)

URL: https://github.com/rapidsai/cuvs/pull/1215