[Java] Support row strides in CuVSMatrix (#1345)
This PR adds full support for row strides to `CuVSMatrix`.

There was already some support for them in `CuVSDeviceMatrix`, as we needed to at least to save them to handle matrices built from tensors (e.g. resulting from a cuvs computation) and/or to create tensors for that kind of data (again, to pass them to the C API).
With this PR we introduce ways to:
- allocate host and device matrices with row stride != from the number of columns (e.g. padding/additional info at the end of row vectors)
- map a pre-allocated memory region to a `CuVSMatrix` taking row stride into account
- copy between different matrices with different strides
- populate data taking the target matrix stride into account
- read data from matrices (e.g. read rows or copy to heap arrays) taking the source matrix stride into account

The API changes include also the introduction of column strides ("gaps" between elements withing the same vector); however this is not implemented yet. To the best of my knowledge, cuvs functions do not support these _yet_. I wanted to introduce them as we need them in some scenarios (e.g. round-tripping from `DLTensor`s), and I did not want to change the API once again if/when we are going to support them.

However, I am open to drop `columnStride` if we do not want it. Let me know.

Authors:
  - Lorenzo Dematté (https://github.com/ldematte)
  - Ben Frederickson (https://github.com/benfred)
  - MithunR (https://github.com/mythrocks)

Approvers:
  - Chris Hegarty (https://github.com/ChrisHegarty)
  - MithunR (https://github.com/mythrocks)

URL: https://github.com/rapidsai/cuvs/pull/1345