Adding more tests and I find the compression settings are broken in the Golang API (#1472)
CreateCompressionParams in index_params.go currently does not work. It always throws an error ""memory allocation failed". This PR fixes that issue and adds a bunch more tests, both to some of the specific files and to cagra_test.go.

I noted that enabling the compression params in cagra_test.go makes the test values not deterministic even if I set the seed. Is this intentional? I looked around online and did not spot a way to set the seed in cuVS. At first I thought it was some C++ variable being not initalized, but added tests verifying all the settings in the compression params...

Thanks

Authors:
  - Max Buckley (https://github.com/maxwbuckley)

Approvers:
  - Ben Frederickson (https://github.com/benfred)

URL: https://github.com/rapidsai/cuvs/pull/1472