fixed bug with update-version.sh (#1556)
Admin merge as the CI failures were unrelated to the changes.

This addresses a bug in the `update-versions.sh` script to tighten the
regular expression to limit the `main` -> `release/YY.MM` pattern so
it's limited to only the `cuvs` URL in the SPDX header section.

* no `non-cuvs` SPDX headers are modified
* all other references to main in the
`python/cuvs_bench/cuvs_bench/plot/__main__.py` file should no longer be
modified

In theory this should forward merge and NOT be reverted as it:
* fixes the `update-versions.sh` script to have limited `regex` 
* reverts the `non-cuvs` urls to use `main` instead of `release/25.12`
<-- will be a no-op in `main`?

---------

Co-authored-by: Bradley Dice <bdice@bradleydice.com>
Co-authored-by: Paul Taylor <178183+trxcllnt@users.noreply.github.com>