Fix BruteForce serialize test (#1568)
`test_serialization.py::test_save_load_brute_force` can sometimes fail due to the ordering of the neighbors. To fix this we can sort the neighbors before comparison, and if there's a small discrepancy due to two neighbors being at the same distance on the `k` cutoff we check that the recall is over 99.8%.

I ran this test on a loop for 60k iterations without it failing.

Authors:
  - Micka (https://github.com/lowener)

Approvers:
  - Gil Forsyth (https://github.com/gforsyth)
  - Divye Gala (https://github.com/divyegala)
  - Corey J. Nolet (https://github.com/cjnolet)

URL: https://github.com/rapidsai/cuvs/pull/1568